move test spec to separate file
