+++ node/src/chain_spec.rs
@@ -203,56 +203,6 @@ pub fn get_chain_spec(para_id: ParaId) -> ChainSpec {
 	)
 }
 
-/// Generate testing chain_spec for staking integration tests with accounts initialized for
-/// collating and nominating.
-#[cfg(feature = "test-staking")]
-pub fn staking_spec(para_id: ParaId) -> ChainSpec {
-	ChainSpec::from_genesis(
-		"Moonbase Development Testnet",
-		"staking",
-		ChainType::Local,
-		move || {
-			testnet_genesis(
-				// Root
-				AccountId::from_str("6Be02d1d3665660d22FF9624b7BE0551ee1Ac91b").unwrap(),
-				// Collators
-				vec![
-					(
-						AccountId::from_str("6Be02d1d3665660d22FF9624b7BE0551ee1Ac91b").unwrap(),
-						None,
-						1_000 * GLMR,
-					),
-					(
-						AccountId::from_str("C0F0f4ab324C46e55D02D0033343B4Be8A55532d").unwrap(),
-						None,
-						1_000 * GLMR,
-					),
-				],
-				// Inflation config
-				moonbeam_inflation_config(),
-				// Endowed accounts (each minted 1 << 80 balance)
-				vec![
-					AccountId::from_str("6Be02d1d3665660d22FF9624b7BE0551ee1Ac91b").unwrap(),
-					AccountId::from_str("C0F0f4ab324C46e55D02D0033343B4Be8A55532d").unwrap(),
-					AccountId::from_str("Ff64d3F6efE2317EE2807d223a0Bdc4c0c49dfDB").unwrap(),
-					AccountId::from_str("f24FF3a9CF04c71Dbc94D0b566f7A27B94566cac").unwrap(),
-				],
-				para_id,
-				// Chain ID
-				1280,
-			)
-		},
-		vec![],
-		None,
-		None,
-		Some(serde_json::from_str("{\"tokenDecimals\": 18}").expect("Provided valid json map")),
-		Extensions {
-			relay_chain: "local_testnet".into(),
-			para_id: para_id.into(),
-		},
-	)
-}
-
 pub fn moonbeam_inflation_config() -> InflationInfo<Balance> {
 	InflationInfo {
 		expect: Range {
@@ -269,7 +219,7 @@ pub fn moonbeam_inflation_config() -> InflationInfo<Balance> {
 	}
 }
 
-fn testnet_genesis(
+pub fn testnet_genesis(
 	root_key: AccountId,
 	stakers: Vec<(AccountId, Option<AccountId>, Balance)>,
 	inflation_config: InflationInfo<Balance>,
